requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
tesseract
transformers
pillow
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
